requirements.txt
litellm>=1.0.0
click>=8.0.0
numpy>=1.20.0
rich>=13.0.0
python-dotenv>=1.0.0
